refactor: edit saving
* move navigation out of component library into frontend
* type the requests
