+++ packages/model/src/user.ts
@@ -100,4 +100,21 @@ export interface UserResponse extends Invitee {
   role: Role;
 }
 
+export interface UserPatchRequest {
+  displayName?: string;
+  email?: string;
+  firstName?: string;
+  lastName?: string;
+  biography?: string;
+  jobTitle?: string;
+  orcid?: string;
+  department?: string;
+  institution?: string;
+  degree?: UserDegree;
+  location?: string;
+  skills?: string[];
+  skillsDescription?: string;
+  questions?: { question: string }[];
+}
+
 export type ListUserResponse = ListResponse<UserResponse>;
